♻️ refactor: restructure project with hybrid domain-driven architecture
- Reviewed MVC and pure domain-driven structures to determine the optimal architecture
- Integrated columns and tasks under the board domain to resolve tight coupling
- Modularized event handlers, rendering logic, and state management for better maintainability
- Introduced a shared folder to manage common utilities and components efficiently
- Refactored folder structure to improve code cohesion and scalability
