+++ index.html
@@ -83,5 +83,5 @@
       </div>
     </dialog>
   </body>
-  <script type="module" src="./js/main.js"></script>
+  <script type="module" src="./src/main.js"></script>
 </html>
