--- js/main.js
@@ -1,5 +1,0 @@
-import { initColumnAndTask } from './components/taskColumn.js';
-import { initActivityPanel } from './components/activity.js';
-
-initColumnAndTask();
-initActivityPanel();
