--- js/utils/fetch.js
@@ -1,14 +1,0 @@
-//get data
-export async function fetchData(url) {
-  try {
-    const response = await fetch(url);
-    if (!response.ok) {
-      throw new Error('네트워크 응답에 문제가 있습니다.');
-    }
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error('Error:', error);
-    return null;
-  }
-}
