+++ src/board/renderers/template.js
@@ -66,27 +66,4 @@ function createTaskCard(id, title, content, author) {
 `;
 }
 
-function createActivityRecord(username, profileImage, actionText, timestamp) {
-  return `  
-    <article class="activity-record">
-      <div class="record-header">
-        <img
-          src="${profileImage}"
-          alt="사용자 프로필 이미지"
-          class="profile-img"
-        />
-      </div>
-      <section class="content">
-        <h2 class="username tx-m14 text-default">@${username}</h2>
-        <p class="action-text tx-m14 text-default">
-          ${actionText}
-        </p>
-        <p class="timestamp tx-m12 text-weak">
-          <time>${timestamp} 전</time>
-        </p>
-      </section>
-    </article>
-  `;
-}
-
-export { createColumn, createTaskCard, createActivityRecord };
+export { createColumn, createTaskCard };
