+++ src/main.js
@@ -1,0 +1,5 @@
+import { initializeBoard } from './board/index.js';
+import { initializeActivity } from './activity/index.js';
+
+initializeBoard();
+initializeActivity();
