+++ src/shared/utils/index.js
@@ -1,0 +1,3 @@
+export * from './dom.js';
+export * from './fetch.js';
+export * from './common.js';
